// src/types/socket.ts
export enum SocketAction {
  UPDATE_APPEARANCE = 'UPDATE_APPEARANCE',

  ADD_COMMENT = 'ADD_COMMENT',
  DELETE_COMMENT = 'DELETE_COMMENT',
  UPDATE_COMMENT = 'UPDATE_COMMENT',

  ADD_IDEA = 'ADD_IDEA',
  DELETE_IDEA = 'DELETE_IDEA',
  UPDATE_IDEA = 'UPDATE_IDEA',

  UPDATE_MODERATION = 'UPDATE_MODERATION',

  UPDATE_NOTIFICATION = 'UPDATE_NOTIFICATION',

  ADD_ROADMAP = 'ADD_ROADMAP',
  DELETE_ROADMAP = 'DELETE_ROADMAP',
  SET_ROADMAPS = 'SET_ROADMAPS',
  UPDATE_ROADMAP = 'UPDATE_ROADMAP',

  DELETE_TAG = 'DELETE_TAG',
  SET_TAGS = 'SET_TAGS',
  UPDATE_TAG = 'UPDATE_TAG',

  ADD_WIDGET = 'ADD_WIDGET',
  DELETE_WIDGET = 'DELETE_WIDGET',
  UPDATE_WIDGET = 'UPDATE_WIDGET',
}